requirements.txt
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-dialogflow==2.0.0
python-telegram-bot==11.1.0
requests==2.25.1
vk-api==11.9.1